Sync test: close to working with a mock sync server
Add a MockSyncServer class that matches the interface used in the main Sync
class. For the moment, the only new test is for the mock sync server itself,
since this implementation doesn't yet terminate the server thread when a
test finishes. Multiple tests fail with 'Address already in use' exceptions.